refactor(agent): connec.ts -> connect.ts
